3. You mustn't be afraid to dream a little bigger, darling

+++ 0x08-recursion/3-factorial.c
@@ -6,7 +6,7 @@
  * return: the factorial of n
  */
 
-int factorial(int n)
+unsigned long long int factorial(unsigned int n)
 {
 	if (n < 0)
 		return (-1);
